requirements.txt
flask
tensorflow-cpu
gunicorn
pillow
numpy
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
imageio